% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bader_apr1_pathways-data.R
\docType{data}
\name{bader_apr1_pathways}
\alias{bader_apr1_pathways}
\title{Pathway table saved from 'pathways_to_test_062817_sorte.txt}
\format{A data table with the transcription start/stop point for >48,000 genes.}
\usage{
data(bader_apr1_pathways)
}
\description{
Pathway table saved from 'pathways_to_test_062817_sorte.txt
}
\examples{
data(bader_apr1_pathways)
print(bader_apr1_pathways[1, ])
}
\keyword{datasets}

