% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_1000G_raw.R
\name{clean_1000G_raw}
\alias{clean_1000G_raw}
\title{clean_1000G_raw.R}
\usage{
clean_1000G_raw(fname_root, gene_name, start_bp, end_bp, checkpoint)
}
\arguments{
\item{fname_root}{The root (everything before .ped) of the downloaded 1000G file.}

\item{gene_name}{A name (often a gene name, but not always) for the region of SNPs you have downloaded,
only to print with error messages.}

\item{start_bp}{The starting position of the range we want.}

\item{end_bp}{The ending position of the range we want.}

\item{checkpoint}{A boolean, if TRUE, print out diagnostic/error messages.}
}
\value{
A list with the elements info_file and ped_file, containing the (clean) genotype info and counts.
}
\description{
After querying data from 1000G, clean the raw data (mostly remove non-biallelic SNPs).
}
\examples{
query_1000G(CHR=1, start_bp=100000, end_bp=200000, buffer=10000, pop_vec='GBR')
}

