% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_info-data.R
\docType{data}
\name{gene_info}
\alias{gene_info}
\title{Hg19 Gene table using mostly Ensembl annotations}
\format{A data table with the transcription start/stop point for >48,000 genes.}
\usage{
data(gene_info)
}
\description{
See make_ensembl_gene_list.R for information on how we created this table.
}
\examples{
data(gene_info)
FGFR2_row <- which(gene_info$HGNC_name == 'FGFR2')
print(gene_info[FGFR2_row, ])
}
\keyword{datasets}

