% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCs_1000G_Euro-data.R
\docType{data}
\name{PCs_1000G_Euro}
\alias{PCs_1000G_Euro}
\title{PCs for 1000G population of 350 European subjects}
\format{A table with first column as ID and rest of columns as PC1-PC20}
\usage{
data(PCs_1000G_Euro)
}
\description{
PCs for 1000G population of 350 European subjects
}
\examples{
data(PCs_1000G_Euro)
print(PCs_1000G_Euro[1, ])
}
\keyword{datasets}

