% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_1000G.R
\name{query_1000G}
\alias{query_1000G}
\title{query_1000G.R}
\usage{
query_1000G(CHR, start_bp, end_bp, gene_name, buffer, pop_vec, Snum, aID,
  checkpoint = T)
}
\arguments{
\item{CHR}{The chromosome where your region is located.}

\item{start_bp}{The starting BP of the region.}

\item{end_bp}{The ending BP of the region.}

\item{gene_name}{Name of region you are querying, for error messages.}

\item{buffer}{Add additional distance to the start and end of the region, possibly to capture nearby
effects such as cis-eQTLs.}

\item{pop_vec}{A vector containing strings describing the populations you want to query. See 1000G
websites for the population codes.}

\item{Snum}{Used in cluster job submission scripts to organize jobs. Must be numeric.}

\item{aID}{Used in cluster job submission scripts to organize jobs. Must be numeric.}

\item{checkpoint}{Would you like the function to print diagnostic/error messages?}
}
\value{
Return 0 for success and 1 for error.
}
\description{
Get 1000G subject SNPs at specific locations on the genome. This is a little less disk-space
intensive than using 'wget' to download entire chromosome files. Only use for one region
at a time (region is one contiguous block on one chromosome).
NOTE: You must have the vcf_to_ped_covert.pl script (provided on 1000G website) in your working
directory!
}
\examples{
query_1000G(CHR=1, start_bp=100000, end_bp=200000, buffer=10000, pop_vec='GBR')
}

