% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{run_tests.R}
\usage{
run_tests(snp_mat, combined_Gmat_record, P_mat = P_mat, run_GHC = FALSE,
  checkpoint = TRUE)
}
\arguments{
\item{snp_mat}{This is the n*p matrix of genotypes from your reference panel. Hopefully the
temp_Gmat variable from prune_snps().}

\item{combined_Gmat_record}{The snp information data frame, hopefully the one output from prune_snps().}

\item{P_mat}{Projection matrix for estimation of covariance matrix.}

\item{run_GHC}{Boolean, if true run both GBJ and GHC, otherwise just GBJ.}

\item{checkpoint}{Boolean, true if you would like to print diagnostic information.}
}
\value{
A list with the elements GBJ, GBJ_p, GBJ_err, GHC, GHC_p, GHC_err, and num_large_Z (number of test
statistics greater than 8.2 in magnitude, which R cannot deal with well).
}
\description{
Run GHC, GBJ, and minP on the summmary statistics.
}

